recordbook api improvements

+++ src/main/java/api/myitmo/MyItmoApi.java
@@ -38,10 +38,10 @@ public interface MyItmoApi {
     Call<ResultResponse<List<Specialization>>> getSpecializations();
 
     @GET("/api/record_book/{specialization_id}/{semester}")
-    Call<ResultResponse<List<RecordBookEntry>>> getRecordBook(@Path("specialization_id") int specializationId, @Path("semester") int semester);
+    Call<ResultResponse<List<RecordBookEntry>>> getRecordBook(@Path("specialization_id") long specializationId, @Path("semester") int semester);
 
     @GET("/api/record_book/{record_book_entry_id}")
-    Call<ResultResponse<List<ControlEntry>>> getControlEntries(@Path("record_book_entry_id") int recordBookEntryId);
+    Call<ResultResponse<List<ControlEntry>>> getControlEntries(@Path("record_book_entry_id") long recordBookEntryId);
 
     // endregion record boot
